moved challenge requests to their own file
